+++ keyboards/helix/rev3/config.h
@@ -1,0 +1,5 @@
+// Copyright 2025 QMK
+// SPDX-License-Identifier: GPL-2.0-or-later
+#pragma once
+
+#define OLED_FONT_H "keyboards/helix/common/glcdfont.c"
