+++ keyboards/helix/rev3/keymaps/default/keymap.json
@@ -1,0 +1,38 @@
+{
+    "keyboard": "helix/rev3",
+    "keymap": "default",
+    "layout": "LAYOUT_5row",
+    "config": {
+        "features": {
+            "encoder_map": true
+        }
+    },
+    "layers": [
+        [
+            "KC_GRV",  "KC_1",    "KC_2",    "KC_3",    "KC_4",         "KC_5",                        "KC_6",    "KC_7",         "KC_8",    "KC_9",    "KC_0",    "KC_DEL",
+            "KC_TAB",  "KC_Q",    "KC_W",    "KC_E",    "KC_R",         "KC_T",                        "KC_Y",    "KC_U",         "KC_I",    "KC_O",    "KC_P",    "KC_BSPC",
+            "KC_LCTL", "KC_A",    "KC_S",    "KC_D",    "KC_F",         "KC_G",                        "KC_H",    "KC_J",         "KC_K",    "KC_L",    "KC_SCLN", "KC_QUOT",
+            "KC_LSFT", "KC_Z",    "KC_X",    "KC_C",    "KC_V",         "KC_B",  "KC_LBRC", "KC_RBRC", "KC_N",    "KC_M",         "KC_COMM", "KC_DOT",  "KC_SLSH", "KC_ENT" ,
+            "MO(3)",   "KC_ESC",  "KC_LALT", "KC_LGUI", "LALT(KC_GRV)", "MO(1)", "KC_SPC",  "KC_SPC",  "MO(2)",   "LALT(KC_GRV)", "KC_LEFT", "KC_DOWN", "KC_UP",   "KC_RGHT"
+        ],
+        [
+            "_______", "_______", "_______", "_______", "_______", "_______",                       "_______", "_______", "_______", "_______", "_______", "_______",
+            "KC_GRV",  "KC_1",    "KC_2",    "KC_3",    "KC_4",    "KC_5",                          "KC_6",    "KC_7",    "KC_8",    "KC_9",    "KC_0",    "KC_DEL",
+            "_______", "KC_F1",   "KC_F2",   "KC_F3",   "KC_F4",   "KC_F5",                         "KC_F6",   "KC_MINS", "KC_EQL",  "KC_LBRC", "KC_RBRC", "KC_BSLS",
+            "KC_CAPS", "KC_F7",   "KC_F8",   "KC_F9",   "KC_F10",  "KC_F11",  "_______", "_______", "KC_F12",  "_______", "_______", "KC_PGDN", "KC_PGUP", "_______",
+            "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "KC_MNXT", "KC_VOLD", "KC_VOLU", "KC_MPLY"
+        ],
+        [
+            "_______", "_______", "_______", "_______", "_______", "_______",                       "_______", "_______", "_______", "_______", "_______", "_______",
+            "KC_F1",   "KC_F2",   "KC_F3",   "KC_F4",   "KC_F5",   "KC_F6",                         "KC_F7",   "KC_F8",   "KC_F9",   "KC_F10",  "KC_F11",  "KC_F12",
+            "QK_BOOT", "_______", "_______", "_______", "_______", "_______",                       "_______", "_______", "_______", "_______", "_______", "_______",
+            "EE_CLR",  "_______", "_______", "_______", "_______", "_______", "AG_NORM", "AG_SWAP", "_______", "_______", "RM_TOGG", "RM_HUEU", "RM_SATU", "RM_VALU",
+            "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "_______", "RM_NEXT", "RM_HUED", "RM_SATD", "RM_VALD"
+        ]
+    ],
+    "encoders": [
+        [{"ccw": "KC_VOLD", "cw": "KC_VOLU"}, {"ccw": "KC_PGUP", "cw": "KC_PGDN"}],
+        [{"ccw": "_______", "cw": "_______"}, {"ccw": "_______", "cw": "_______"}],
+        [{"ccw": "_______", "cw": "_______"}, {"ccw": "_______", "cw": "_______"}]
+    ]
+}
