--- keyboards/helix/rev3_4rows/keyboard.json
@@ -1,118 +1,0 @@
-{
-  "keyboard_name": "Helix rev3 4rows",
-  "manufacturer": "yushakobo",
-  "maintainer": "yushakobo",
-  "usb": {
-    "vid": "0x3265",
-    "pid": "0x0004",
-    "device_version": "0.0.1"
-  },
-  "qmk": {
-    "locking": {
-      "enabled": true,
-      "resync": true
-    }
-  },
-  "rgb_matrix": {
-    "driver": "ws2812",
-    "sat_steps": 8,
-    "val_steps": 8,
-    "speed_steps": 10,
-    "max_brightness": 150,
-    "split_count": [25, 25],
-    "sleep": true
-  },
-  "dip_switch": {
-    "matrix_grid": [ [0,6], [1,6], [5,6], [6,6] ]
-  },
-  "encoder": {
-    "rotary": [
-      {"pin_a": "B6", "pin_b": "B5"}
-    ]
-  },
-  "split": {
-    "enabled": true,
-    "serial": {
-        "pin": "D2"
-     }
-  },
-  "ws2812": {
-    "pin": "D3"
-  },
-  "rgblight": {
-    "led_count": 50,
-    "max_brightness": 120,
-    "split_count": [25, 25],
-    "animations": {
-      "rainbow_mood": true,
-      "rainbow_swirl": true,
-      "static_gradient": true,
-      "rgb_test": true
-    }
-  },
-  "processor": "atmega32u4",
-  "bootloader": "caterina",
-  "layouts": {
-    "LAYOUT": {
-      "layout": [
-        {"matrix": [0, 0], "x": 0, "y": 0},
-        {"matrix": [0, 1], "x": 1, "y": 0},
-        {"matrix": [0, 2], "x": 2, "y": 0},
-        {"matrix": [0, 3], "x": 3, "y": 0},
-        {"matrix": [0, 4], "x": 4, "y": 0},
-        {"matrix": [0, 5], "x": 5, "y": 0},
-
-        {"matrix": [4, 5], "x": 9, "y": 0},
-        {"matrix": [4, 4], "x": 10, "y": 0},
-        {"matrix": [4, 3], "x": 11, "y": 0},
-        {"matrix": [4, 2], "x": 12, "y": 0},
-        {"matrix": [4, 1], "x": 13, "y": 0},
-        {"matrix": [4, 0], "x": 14, "y": 0},
-
-        {"matrix": [1, 0], "x": 0, "y": 1},
-        {"matrix": [1, 1], "x": 1, "y": 1},
-        {"matrix": [1, 2], "x": 2, "y": 1},
-        {"matrix": [1, 3], "x": 3, "y": 1},
-        {"matrix": [1, 4], "x": 4, "y": 1},
-        {"matrix": [1, 5], "x": 5, "y": 1},
-
-        {"matrix": [5, 5], "x": 9, "y": 1},
-        {"matrix": [5, 4], "x": 10, "y": 1},
-        {"matrix": [5, 3], "x": 11, "y": 1},
-        {"matrix": [5, 2], "x": 12, "y": 1},
-        {"matrix": [5, 1], "x": 13, "y": 1},
-        {"matrix": [5, 0], "x": 14, "y": 1},
-
-        {"matrix": [2, 0], "x": 0, "y": 2},
-        {"matrix": [2, 1], "x": 1, "y": 2},
-        {"matrix": [2, 2], "x": 2, "y": 2},
-        {"matrix": [2, 3], "x": 3, "y": 2},
-        {"matrix": [2, 4], "x": 4, "y": 2},
-        {"matrix": [2, 5], "x": 5, "y": 2},
-
-        {"matrix": [6, 5], "x": 9, "y": 2},
-        {"matrix": [6, 4], "x": 10, "y": 2},
-        {"matrix": [6, 3], "x": 11, "y": 2},
-        {"matrix": [6, 2], "x": 12, "y": 2},
-        {"matrix": [6, 1], "x": 13, "y": 2},
-        {"matrix": [6, 0], "x": 14, "y": 2},
-
-        {"matrix": [3, 0], "x": 0, "y": 3},
-        {"matrix": [3, 1], "x": 1, "y": 3},
-        {"matrix": [3, 2], "x": 2, "y": 3},
-        {"matrix": [3, 3], "x": 3, "y": 3},
-        {"matrix": [3, 4], "x": 4, "y": 3},
-        {"matrix": [3, 5], "x": 5, "y": 3},
-        {"matrix": [3, 6], "x": 6, "y": 3},
-
-        {"matrix": [7, 6], "x": 8, "y": 3},
-        {"matrix": [7, 5], "x": 9, "y": 3},
-        {"matrix": [7, 4], "x": 10, "y": 3},
-        {"matrix": [7, 3], "x": 11, "y": 3},
-        {"matrix": [7, 2], "x": 12, "y": 3},
-        {"matrix": [7, 1], "x": 13, "y": 3},
-        {"matrix": [7, 0], "x": 14, "y": 3}
-      ]
-    }
-  }
-}
