--- keyboards/helix/rev3_4rows/keymaps/default/readme.md
@@ -1,66 +1,0 @@
-# The default keymap for Helix rev3 4rows
-
-### Qwerty Layer (Base)
-```
-,-----------------------------------------.             ,-----------------------------------------.
-| Tab  |   Q  |   W  |   E  |   R  |   T  |             |   Y  |   U  |   I  |   O  |   P  | Bksp |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-| Ctrl |   A  |   S  |   D  |   F  |   G  |             |   H  |   J  |   K  |   L  |   ;  |  '   |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-| Shift|   Z  |   X  |   C  |   V  |   B  |             |   N  |   M  |   ,  |   .  |   /  |Enter |
-|------+------+------+------+------+------+-------------+------+------+------+------+------+------|
-|Adjust| Esc  | Alt  | GUI  | EISU |Lower |Space |Space |Raise | KANA | Left | Down |  Up  |Right |
-`-------------------------------------------------------------------------------------------------'
-```
-### Lower Layer
-```
-,-----------------------------------------.             ,-----------------------------------------.
-|   ~  |   !  |   @  |   #  |   $  |   %  |             |   ^  |   &  |   *  |   (  |   )  |      |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-|      |  F1  |  F2  |  F3  |  F4  |  F5  |             |  F6  |   _  |   +  |   {  |   }  |  |   |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-| CAPS |  F7  |  F8  |  F9  |  F10 |  F11 |             |  F12 |      |      | Home | End  |      |
-|------+------+------+------+------+------+-------------+------+------+------+------+------+------|
-|      |      |      |      |      |      |      |      |      |      | Next | Vol- | Vol+ | Play |
-`-------------------------------------------------------------------------------------------------'
-```
-### Raise Layer
-```
-,-----------------------------------------.             ,-----------------------------------------.
-|   `  |   1  |   2  |   3  |   4  |   5  |             |   6  |   7  |   8  |   9  |   0  | Del  |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-|      |  F1  |  F2  |  F3  |  F4  |  F5  |             |  F6  |   -  |   =  |   [  |   ]  |  \   |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-| CAPS |  F7  |  F8  |  F9  |  F10 |  F11 |             |  F12 |      |      |PageDn|PageUp|      |
-|------+------+------+------+------+------+-------------+------+------+------+------+------+------|
-|      |      |      |      |      |      |      |      |      |      | Next | Vol- | Vol+ | Play |
-`-------------------------------------------------------------------------------------------------'
-```
-### Adjust Layer (Lower + Raise)
-```
-,-----------------------------------------.             ,-----------------------------------------.
-|  F1  |  F2  |  F3  |  F4  |  F5  |  F6  |             |  F7  |  F8  |  F9  |  F10 |  F11 |  F12 |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-|      | Reset|RGBRST|EEPRST|      |      |             |      |      |      |      |      |  Del |
-|------+------+------+------+------+------|             |------+------+------+------+------+------|
-|      |      |      |      |      | Mac  |             | Win  |      |RGB ON| HUE+ | SAT+ | VAL+ |
-|------+------+------+------+------+------+------+------+------+------+------+------+------+------|
-|      |      |      |      |      |      |      |      |      |      | MODE | HUE- | SAT- | VAL- |
-`-------------------------------------------------------------------------------------------------'
-```
-
-### Rotary Encoder Settings
-
-|L/R|Clockwise|Counter Clockwise|
-|---|---|---|
-|Left|Page Down|Page Up|
-|Right|Down|Up|
-
-### DipSW Settings
-
-|Switch No.||
-|---|---|
-|Left 1|Toggle Win/mac|
-|Left 2|Not assigned|
-|Right 1|Not assigned|
-|Right 2|Not assigned|
