--- keyboards/helix/rev3_4rows/rules.mk
@@ -1,9 +1,0 @@
-EXTRAKEY_ENABLE = yes    # Audio control and System control
-RGBLIGHT_ENABLE = yes   # Enable keyboard RGB underglow
-RGB_MATRIX_ENABLE = no
-OLED_ENABLE = yes
-ENCODER_ENABLE = yes
-DIP_SWITCH_ENABLE = no
-LTO_ENABLE = yes
-
-SRC += oled_display.c
