--- keyboards/helix/rev3_5rows/keyboard.json
@@ -1,198 +1,0 @@
-{
-  "keyboard_name": "Helix rev3 5rows",
-  "manufacturer": "yushakobo",
-  "maintainer": "yushakobo",
-  "usb": {
-    "vid": "0x3265",
-    "pid": "0x0003",
-    "device_version": "0.0.1"
-  },
-  "qmk": {
-    "locking": {
-      "enabled": true,
-      "resync": true
-    }
-  },
-  "rgb_matrix": {
-    "driver": "ws2812",
-    "sat_steps": 8,
-    "val_steps": 8,
-    "speed_steps": 10,
-    "layout": [
-        {"matrix": [0, 5], "x": 80, "y": 0, "flags": 4},
-        {"matrix": [0, 4], "x": 64, "y": 0, "flags": 4},
-        {"matrix": [0, 3], "x": 48, "y": 0, "flags": 4},
-        {"matrix": [0, 2], "x": 32, "y": 0, "flags": 4},
-        {"matrix": [0, 1], "x": 16, "y": 0, "flags": 4},
-        {"matrix": [0, 0], "x": 0, "y": 0, "flags": 4},
-        {"matrix": [1, 0], "x": 0, "y": 16, "flags": 4},
-        {"matrix": [1, 1], "x": 16, "y": 16, "flags": 4},
-        {"matrix": [1, 2], "x": 32, "y": 16, "flags": 4},
-        {"matrix": [1, 3], "x": 48, "y": 16, "flags": 4},
-        {"matrix": [1, 4], "x": 64, "y": 16, "flags": 4},
-        {"matrix": [1, 5], "x": 80, "y": 16, "flags": 4},
-        {"matrix": [2, 5], "x": 80, "y": 32, "flags": 4},
-        {"matrix": [2, 4], "x": 64, "y": 32, "flags": 4},
-        {"matrix": [2, 3], "x": 48, "y": 32, "flags": 4},
-        {"matrix": [2, 2], "x": 32, "y": 32, "flags": 4},
-        {"matrix": [2, 1], "x": 16, "y": 32, "flags": 4},
-        {"matrix": [2, 0], "x": 0, "y": 32, "flags": 4},
-        {"matrix": [3, 0], "x": 0, "y": 48, "flags": 4},
-        {"matrix": [3, 1], "x": 16, "y": 48, "flags": 4},
-        {"matrix": [3, 2], "x": 32, "y": 48, "flags": 4},
-        {"matrix": [3, 3], "x": 48, "y": 48, "flags": 4},
-        {"matrix": [3, 4], "x": 64, "y": 48, "flags": 4},
-        {"matrix": [3, 5], "x": 80, "y": 48, "flags": 4},
-        {"matrix": [3, 6], "x": 96, "y": 48, "flags": 4},
-        {"matrix": [4, 6], "x": 96, "y": 64, "flags": 4},
-        {"matrix": [4, 5], "x": 80, "y": 64, "flags": 4},
-        {"matrix": [4, 4], "x": 64, "y": 64, "flags": 4},
-        {"matrix": [4, 3], "x": 48, "y": 64, "flags": 4},
-        {"matrix": [4, 2], "x": 32, "y": 64, "flags": 4},
-        {"matrix": [4, 1], "x": 16, "y": 64, "flags": 4},
-        {"matrix": [4, 0], "x": 0, "y": 64, "flags": 4},
-        {"matrix": [5, 5], "x": 144, "y": 0, "flags": 4},
-        {"matrix": [5, 4], "x": 160, "y": 0, "flags": 4},
-        {"matrix": [5, 3], "x": 176, "y": 0, "flags": 4},
-        {"matrix": [5, 2], "x": 192, "y": 0, "flags": 4},
-        {"matrix": [5, 1], "x": 208, "y": 0, "flags": 4},
-        {"matrix": [5, 0], "x": 224, "y": 0, "flags": 4},
-        {"matrix": [6, 0], "x": 224, "y": 16, "flags": 4},
-        {"matrix": [6, 1], "x": 208, "y": 16, "flags": 4},
-        {"matrix": [6, 2], "x": 192, "y": 16, "flags": 4},
-        {"matrix": [6, 3], "x": 176, "y": 16, "flags": 4},
-        {"matrix": [6, 4], "x": 160, "y": 16, "flags": 4},
-        {"matrix": [6, 5], "x": 144, "y": 16, "flags": 4},
-        {"matrix": [7, 5], "x": 144, "y": 32, "flags": 4},
-        {"matrix": [7, 4], "x": 160, "y": 32, "flags": 4},
-        {"matrix": [7, 3], "x": 176, "y": 32, "flags": 4},
-        {"matrix": [7, 2], "x": 192, "y": 32, "flags": 4},
-        {"matrix": [7, 1], "x": 208, "y": 32, "flags": 4},
-        {"matrix": [7, 0], "x": 224, "y": 32, "flags": 4},
-        {"matrix": [8, 0], "x": 224, "y": 48, "flags": 4},
-        {"matrix": [8, 1], "x": 208, "y": 48, "flags": 4},
-        {"matrix": [8, 2], "x": 192, "y": 48, "flags": 4},
-        {"matrix": [8, 3], "x": 176, "y": 48, "flags": 4},
-        {"matrix": [8, 4], "x": 160, "y": 48, "flags": 4},
-        {"matrix": [8, 5], "x": 144, "y": 48, "flags": 4},
-        {"matrix": [8, 6], "x": 128, "y": 48, "flags": 4},
-        {"matrix": [9, 6], "x": 128, "y": 64, "flags": 4},
-        {"matrix": [9, 5], "x": 144, "y": 64, "flags": 4},
-        {"matrix": [9, 4], "x": 160, "y": 64, "flags": 4},
-        {"matrix": [9, 3], "x": 176, "y": 64, "flags": 4},
-        {"matrix": [9, 2], "x": 192, "y": 64, "flags": 4},
-        {"matrix": [9, 1], "x": 208, "y": 64, "flags": 4},
-        {"matrix": [9, 0], "x": 224, "y": 64, "flags": 4}
-    ],
-    "max_brightness": 128,
-    "split_count": [32, 32],
-    "sleep": true
-  },
-  "dip_switch": {
-    "matrix_grid": [ [0,6], [1,6], [5,6], [6,6] ]
-  },
-  "encoder": {
-    "rotary": [
-      {"pin_a": "B6", "pin_b": "B5"}
-    ]
-  },
-  "split": {
-    "enabled": true,
-    "serial": {
-        "pin": "D2"
-    }
-  },
-  "ws2812": {
-    "pin": "D3"
-  },
-  "rgblight": {
-    "led_count": 64,
-    "max_brightness": 120,
-    "split_count": [32, 32],
-    "animations": {
-      "rainbow_mood": true,
-      "rainbow_swirl": true
-    }
-  },
-  "processor": "atmega32u4",
-  "bootloader": "caterina",
-  "layouts": {
-    "LAYOUT": {
-      "layout": [
-        {"matrix": [0, 0], "x": 0, "y": 0},
-        {"matrix": [0, 1], "x": 1, "y": 0},
-        {"matrix": [0, 2], "x": 2, "y": 0},
-        {"matrix": [0, 3], "x": 3, "y": 0},
-        {"matrix": [0, 4], "x": 4, "y": 0},
-        {"matrix": [0, 5], "x": 5, "y": 0},
-
-        {"matrix": [5, 5], "x": 9, "y": 0},
-        {"matrix": [5, 4], "x": 10, "y": 0},
-        {"matrix": [5, 3], "x": 11, "y": 0},
-        {"matrix": [5, 2], "x": 12, "y": 0},
-        {"matrix": [5, 1], "x": 13, "y": 0},
-        {"matrix": [5, 0], "x": 14, "y": 0},
-
-        {"matrix": [1, 0], "x": 0, "y": 1},
-        {"matrix": [1, 1], "x": 1, "y": 1},
-        {"matrix": [1, 2], "x": 2, "y": 1},
-        {"matrix": [1, 3], "x": 3, "y": 1},
-        {"matrix": [1, 4], "x": 4, "y": 1},
-        {"matrix": [1, 5], "x": 5, "y": 1},
-
-        {"matrix": [6, 5], "x": 9, "y": 1},
-        {"matrix": [6, 4], "x": 10, "y": 1},
-        {"matrix": [6, 3], "x": 11, "y": 1},
-        {"matrix": [6, 2], "x": 12, "y": 1},
-        {"matrix": [6, 1], "x": 13, "y": 1},
-        {"matrix": [6, 0], "x": 14, "y": 1},
-
-        {"matrix": [2, 0], "x": 0, "y": 2},
-        {"matrix": [2, 1], "x": 1, "y": 2},
-        {"matrix": [2, 2], "x": 2, "y": 2},
-        {"matrix": [2, 3], "x": 3, "y": 2},
-        {"matrix": [2, 4], "x": 4, "y": 2},
-        {"matrix": [2, 5], "x": 5, "y": 2},
-
-        {"matrix": [7, 5], "x": 9, "y": 2},
-        {"matrix": [7, 4], "x": 10, "y": 2},
-        {"matrix": [7, 3], "x": 11, "y": 2},
-        {"matrix": [7, 2], "x": 12, "y": 2},
-        {"matrix": [7, 1], "x": 13, "y": 2},
-        {"matrix": [7, 0], "x": 14, "y": 2},
-
-        {"matrix": [3, 0], "x": 0, "y": 3},
-        {"matrix": [3, 1], "x": 1, "y": 3},
-        {"matrix": [3, 2], "x": 2, "y": 3},
-        {"matrix": [3, 3], "x": 3, "y": 3},
-        {"matrix": [3, 4], "x": 4, "y": 3},
-        {"matrix": [3, 5], "x": 5, "y": 3},
-        {"matrix": [3, 6], "x": 6, "y": 3},
-
-        {"matrix": [8, 6], "x": 8, "y": 3},
-        {"matrix": [8, 5], "x": 9, "y": 3},
-        {"matrix": [8, 4], "x": 10, "y": 3},
-        {"matrix": [8, 3], "x": 11, "y": 3},
-        {"matrix": [8, 2], "x": 12, "y": 3},
-        {"matrix": [8, 1], "x": 13, "y": 3},
-        {"matrix": [8, 0], "x": 14, "y": 3},
-
-        {"matrix": [4, 0], "x": 0, "y": 4},
-        {"matrix": [4, 1], "x": 1, "y": 4},
-        {"matrix": [4, 2], "x": 2, "y": 4},
-        {"matrix": [4, 3], "x": 3, "y": 4},
-        {"matrix": [4, 4], "x": 4, "y": 4},
-        {"matrix": [4, 5], "x": 5, "y": 4},
-        {"matrix": [4, 6], "x": 6, "y": 4},
-
-        {"matrix": [9, 6], "x": 8, "y": 4},
-        {"matrix": [9, 5], "x": 9, "y": 4},
-        {"matrix": [9, 4], "x": 10, "y": 4},
-        {"matrix": [9, 3], "x": 11, "y": 4},
-        {"matrix": [9, 2], "x": 12, "y": 4},
-        {"matrix": [9, 1], "x": 13, "y": 4},
-        {"matrix": [9, 0], "x": 14, "y": 4}
-      ]
-    }
-  }
-}
