--- keyboards/helix/rev3_5rows/keymaps/default/rules.mk
@@ -1,1 +1,0 @@
-ENCODER_MAP_ENABLE = yes
