--- keyboards/helix/rev3_5rows/oled_display.c
@@ -1,9 +1,0 @@
-/*
-
- There are several ways to create your own keymap oled code:
-
- * Add the oled code to your keymaps/<keymap_name>/keymap.c.
- * Create a new file in your keymaps/<keymap_name>/ directory, add the oled code, and add `SRC + = <filename>` to keymaps/<keymap_name>/rules.mk.
- * Copy keymaps/default/oled_display.c to your keymaps/<keymap_name>/ directory and modify it.
-
-*/
