--- keyboards/helix/rev3_5rows/readme.md
@@ -1,13 +1,0 @@
-# Helix rev3 5rows
-
-A compact split ortholinear keyboard.
-
-* Keyboard Maintainer: [yushakobo](https://github.com/yushakobo)
-* Hardware Supported: Helix rev3 PCBs, Pro Micro
-* Hardware Availability: (Under preparation)
-
-Make example for this keyboard (after setting up your build environment):
-
-    make helix/rev3_5rows:default:flash
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
